# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * COBO Africa Payments Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  merchant: "merchant",
  customer: "customer",
} as const;

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const UserStatus = {
  active: "active",
  suspended: "suspended",
  pending: "pending",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  status: UserStatus;
  phone?: string | null;
  country?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export interface UsersListResponse {
  data: User[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  admin: "admin",
  merchant: "merchant",
  customer: "customer",
} as const;

export interface CreateUserBody {
  email: string;
  name: string;
  password: string;
  role: CreateUserBodyRole;
  phone?: string;
  country?: string;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  admin: "admin",
  merchant: "merchant",
  customer: "customer",
} as const;

export type UpdateUserBodyStatus =
  (typeof UpdateUserBodyStatus)[keyof typeof UpdateUserBodyStatus];

export const UpdateUserBodyStatus = {
  active: "active",
  suspended: "suspended",
  pending: "pending",
} as const;

export interface UpdateUserBody {
  name?: string;
  email?: string;
  role?: UpdateUserBodyRole;
  status?: UpdateUserBodyStatus;
  phone?: string;
  country?: string;
}

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
  refunded: "refunded",
} as const;

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  payment: "payment",
  transfer: "transfer",
  withdrawal: "withdrawal",
  deposit: "deposit",
} as const;

export interface Transaction {
  id: number;
  reference: string;
  amount: number;
  currency: string;
  status: TransactionStatus;
  type: TransactionType;
  merchantId?: number | null;
  merchantName?: string | null;
  customerId?: number | null;
  customerName?: string | null;
  description?: string | null;
  country?: string | null;
  paymentMethod?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface TransactionsListResponse {
  data: Transaction[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type CreateTransactionBodyType =
  (typeof CreateTransactionBodyType)[keyof typeof CreateTransactionBodyType];

export const CreateTransactionBodyType = {
  payment: "payment",
  transfer: "transfer",
  withdrawal: "withdrawal",
  deposit: "deposit",
} as const;

export interface CreateTransactionBody {
  amount: number;
  currency: string;
  type: CreateTransactionBodyType;
  merchantId?: number;
  customerId?: number;
  description?: string;
  country?: string;
  paymentMethod?: string;
}

export type UpdateTransactionBodyStatus =
  (typeof UpdateTransactionBodyStatus)[keyof typeof UpdateTransactionBodyStatus];

export const UpdateTransactionBodyStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
  refunded: "refunded",
} as const;

export interface UpdateTransactionBody {
  status?: UpdateTransactionBodyStatus;
  description?: string;
}

export type MerchantStatus =
  (typeof MerchantStatus)[keyof typeof MerchantStatus];

export const MerchantStatus = {
  active: "active",
  suspended: "suspended",
  pending: "pending",
} as const;

export interface Merchant {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  country: string;
  status: MerchantStatus;
  businessType?: string | null;
  totalVolume?: number | null;
  transactionCount?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface MerchantsListResponse {
  data: Merchant[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateMerchantBody {
  name: string;
  email: string;
  phone?: string;
  country: string;
  businessType?: string;
}

export type UpdateMerchantBodyStatus =
  (typeof UpdateMerchantBodyStatus)[keyof typeof UpdateMerchantBodyStatus];

export const UpdateMerchantBodyStatus = {
  active: "active",
  suspended: "suspended",
  pending: "pending",
} as const;

export interface UpdateMerchantBody {
  name?: string;
  email?: string;
  phone?: string;
  country?: string;
  status?: UpdateMerchantBodyStatus;
  businessType?: string;
}

export interface DashboardSummary {
  totalTransactions: number;
  totalVolume: number;
  successRate: number;
  activeMerchants: number;
  pendingTransactions: number;
  totalUsers: number;
  volumeGrowth: number;
  transactionGrowth: number;
}

export interface CountryVolume {
  country: string;
  volume: number;
  count: number;
}

export interface MonthlyVolume {
  month: string;
  volume: number;
  count: number;
}

export interface MerchantStat {
  id: number;
  name: string;
  country: string;
  volume: number;
  count: number;
}

export type ListUsersParams = {
  page?: number;
  limit?: number;
  search?: string;
  role?: ListUsersRole;
  status?: ListUsersStatus;
};

export type ListUsersRole = (typeof ListUsersRole)[keyof typeof ListUsersRole];

export const ListUsersRole = {
  admin: "admin",
  merchant: "merchant",
  customer: "customer",
} as const;

export type ListUsersStatus =
  (typeof ListUsersStatus)[keyof typeof ListUsersStatus];

export const ListUsersStatus = {
  active: "active",
  suspended: "suspended",
  pending: "pending",
} as const;

export type ListTransactionsParams = {
  page?: number;
  limit?: number;
  status?: ListTransactionsStatus;
  merchantId?: number;
  startDate?: string;
  endDate?: string;
  search?: string;
};

export type ListTransactionsStatus =
  (typeof ListTransactionsStatus)[keyof typeof ListTransactionsStatus];

export const ListTransactionsStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
  refunded: "refunded",
} as const;

export type ListMerchantsParams = {
  page?: number;
  limit?: number;
  search?: string;
  status?: ListMerchantsStatus;
  country?: string;
};

export type ListMerchantsStatus =
  (typeof ListMerchantsStatus)[keyof typeof ListMerchantsStatus];

export const ListMerchantsStatus = {
  active: "active",
  suspended: "suspended",
  pending: "pending",
} as const;

export type GetRecentTransactionsParams = {
  limit?: number;
};

export type GetTopMerchantsParams = {
  limit?: number;
};