# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: COBO Africa Payments Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      description: Authenticate user and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, failed, refunded]
        - name: merchantId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionsListResponse"

    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create transaction
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update transaction status
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionBody"
      responses:
        "200":
          description: Transaction updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /merchants:
    get:
      operationId: listMerchants
      tags: [merchants]
      summary: List merchants
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, pending]
        - name: country
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of merchants
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantsListResponse"

    post:
      operationId: createMerchant
      tags: [merchants]
      summary: Create merchant
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMerchantBody"
      responses:
        "201":
          description: Merchant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"
  /merchants/{id}:
    get:
      operationId: getMerchant
      tags: [merchants]
      summary: Get merchant by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Merchant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"

    put:
      operationId: updateMerchant
      tags: [merchants]
      summary: Update merchant
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMerchantBody"
      responses:
        "200":
          description: Merchant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"

    delete:
      operationId: deleteMerchant
      tags: [merchants]
      summary: Delete merchant
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Merchant deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, merchant, customer]
        status:
          type: string
          enum: [active, suspended, pending]
        phone:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - status
        - createdAt
        - updatedAt
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        reference:
          type: string
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        type:
          type: string
          enum: [payment, transfer, withdrawal, deposit]
        merchantId:
          type: integer
          nullable: true
        merchantName:
          type: string
          nullable: true
        customerId:
          type: integer
          nullable: true
        customerName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        paymentMethod:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - reference
        - amount
        - currency
        - status
        - type
        - createdAt
        - updatedAt
    TransactionsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - totalPages
    CreateTransactionBody:
      type: object
      properties:
        amount:
          type: number
        currency:
          type: string
        type:
          type: string
          enum: [payment, transfer, withdrawal, deposit]
        merchantId:
          type: integer
        customerId:
          type: integer
        description:
          type: string
        country:
          type: string
        paymentMethod:
          type: string
      required:
        - amount
        - currency
        - type
    UpdateTransactionBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        description:
          type: string
    Merchant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        country:
          type: string
        status:
          type: string
          enum: [active, suspended, pending]
        businessType:
          type: string
          nullable: true
        totalVolume:
          type: number
          nullable: true
        transactionCount:
          type: integer
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
        - country
        - status
        - createdAt
        - updatedAt
    MerchantsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Merchant"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - totalPages
    CreateMerchantBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        country:
          type: string
        businessType:
          type: string
      required:
        - name
        - email
        - country
    UpdateMerchantBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        country:
          type: string
        status:
          type: string
          enum: [active, suspended, pending]
        businessType:
          type: string
...