# lib/api-zod/src/generated/types/createMerchantBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * COBO Africa Payments Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMerchantBody {
  name: string;
  email: string;
  phone?: string;
  country: string;
  businessType?: string;
}